pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "ponbac", email = "root@ponbac.xyz"},
]
dependencies = [
    "pydantic==1.9.0",
    "python-dotenv==0.20.0",
    "realtime==0.0.4",
    "supabase==0.5.8",
    "numpy",
    "fastapi>=0.87.0",
    "uvicorn[standard]>=0.15.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]

[tool.pdm.scripts]
start = "uvicorn main:app --reload --host 0.0.0.0 --port 3999"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
